Move Anchor/Order enums to lib.rs
