+++ shards/egui/src/egui_host.rs
@@ -11,7 +11,7 @@ use crate::bindings::make_native_full_output;
 use crate::bindings::NativeFullOutput;
 use shards::core::Core;
 
-use shards::shardsc::Shards;
+use shards::shardsc::Shards; 
 use shards::types::Context;
 use shards::types::ExposedInfo;
